+++ frontend/src/api/renderPanel.ts
@@ -1,0 +1,15 @@
+import * as v1 from "../schemas/v1";
+import { fetchJSON } from "./api";
+
+export interface RenderPanelResponse {
+  panel: v1.Panel;
+}
+
+/** @deprecated */
+export async function renderPanel(panel: v1.Panel) {
+  const params = new URLSearchParams(location.search);
+  const url = `v1_render_panel?${params}`;
+
+  const response = await fetchJSON<RenderPanelResponse>(url, { panel });
+  return response.panel;
+}