+++ frontend/src/panels/ReactPanel.tsx
@@ -1,0 +1,6 @@
+import { ReactElement } from "react";
+import { PanelProps } from "./registry";
+
+export function ReactPanel(_props: PanelProps<ReactElement>) {
+  return <></>;
+}