+++ frontend/src/schemas/v1_helpers.ts
@@ -6,7 +6,7 @@ export const iterateMonths = (dateFirst: string, dateLast: string) => {
 
   const months: { year: number; month: number }[] = [];
   let [year, month] = dateFirst.split("-").map((x) => parseInt(x));
-  let [lastYear, lastMonth] = dateLast.split("-").map((x) => parseInt(x));
+  const [lastYear, lastMonth] = dateLast.split("-").map((x) => parseInt(x));
 
   while (year < lastYear || (year === lastYear && month <= lastMonth)) {
     months.push({ year, month });
@@ -28,7 +28,7 @@ export const iterateYears = (dateFirst: string, dateLast: string) => {
 
   const years: number[] = [];
   let year = parseInt(dateFirst.split("-")[0]);
-  let lastYear = parseInt(dateLast.split("-")[0]);
+  const lastYear = parseInt(dateLast.split("-")[0]);
 
   for (; year <= lastYear; year++) {
     years.push(year);
@@ -46,8 +46,8 @@ interface AccountTreeNode {
  * @deprecated
  */
 export const buildAccountTree = (
-  rows: any[],
-  valueFn: (row: any) => number,
+  rows: any[], // eslint-disable-line @typescript-eslint/no-explicit-any
+  valueFn: (row: any) => number, // eslint-disable-line @typescript-eslint/no-explicit-any
   nameFn: (parts: string[], i: number) => string,
 ) => {
   console.warn("helpers.iterateMonths() is deprecated, please define this function in utils.inline in dashboards.yaml");
@@ -55,7 +55,7 @@ export const buildAccountTree = (
   nameFn = nameFn ?? ((parts: string[], i: number) => parts.slice(0, i + 1).join(":"));
 
   const accountTree: { children: AccountTreeNode[] } = { children: [] };
-  for (let row of rows) {
+  for (const row of rows) {
     const accountParts = row.account.split(":");
     let node = accountTree;
     for (let i = 0; i < accountParts.length; i++) {