+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v2.0.0 (unreleased)
+### Added
+* support dynamic, typed dashboards written in TypeScript (`dashboards.tsx`) or JavaScript (`dashboards.jsx`)
+* concurrent loading of panels
+* new `react` panel kind
+
+### Changed
+* updated design using Material UI
+
+### Removed
+* the `jinja2` panel type is deprecated in TypeScript/JavaScript dashboards
+
+Legacy `dashboards.yaml` configurations will be migrated transparently and are expected to work.
+
 ## v1.2.0 (2025-10-15)
 * upgrade echarts to v6.0.0